Upload platform independent teapot and new configuration script
This update consists of many bux fixes as well as a rewrite of the configuration script. Now, the script does not overstep its boundaries by deleting and moving files -- it simply changes the appropriate lines in 'teapot'. By default, teapot uses ffmpeg, but if the user is on a Mac the configuration script will change that to afplay. I've also changed some key mappings in teapot, and added a PID global variable containing the process id of the song playing. Bugs regarding EXTENSIONS were fixed -- ext_list is no longer in existence, and EXTENSIONS is an array (not relying on any dodgy behavior). The mechanism for num_songs is different (and actually functional). show_music_player's nonsensical code was replaced (now it simply relays messages to show_options, which can either 'reset', set seconds to a value, or just show options with the default amount of seconds. Pausing works but is still dependent on the process not being killed randomly by the system (which means you can't pause for extensive amounts of time). That's about it!